refactor: inject Tenor API key via constructor
