+++ apps/akari/jest.config.js
@@ -20,5 +20,6 @@ module.exports = {
   moduleNameMapper: {
     '^@/bluesky-api$': '<rootDir>/../../packages/bluesky-api/src',
     '^@/clearsky-api$': '<rootDir>/../../packages/clearsky-api/src/api',
+    '^@/tenor-api$': '<rootDir>/../../packages/tenor-api/src',
   },
 };
