+++ apps/akari/metro.config.js
@@ -3,9 +3,10 @@ const path = require('path');
 
 const config = getDefaultConfig(__dirname);
 
-// Add path mapping for the bluesky-api package
+// Add path mapping for the internal packages
 config.resolver.alias = {
   '@/bluesky-api': path.resolve(__dirname, '../../packages/bluesky-api'),
+  '@/tenor-api': path.resolve(__dirname, '../../packages/tenor-api'),
 };
 
 module.exports = config;
