+++ apps/akari/package.json
@@ -36,6 +36,7 @@
     "@tanstack/react-query-persist-client": "^5.83.0",
     "bluesky-api": "*",
     "clearsky-api": "*",
+    "tenor-api": "*",
     "expo": "~53.0.20",
     "expo-blur": "~14.1.5",
     "expo-clipboard": "~7.1.5",
