+++ apps/akari/tsconfig.json
@@ -7,7 +7,8 @@
     "paths": {
       "@/*": ["./*"],
       "@/bluesky-api": ["../../packages/bluesky-api"],
-      "@/clearsky-api": ["../../packages/clearsky-api"]
+      "@/clearsky-api": ["../../packages/clearsky-api"],
+      "@/tenor-api": ["../../packages/tenor-api"]
     }
   },
   "include": ["**/*.ts", "**/*.tsx", "**/*.json", ".expo/types/**/*.ts", "expo-env.d.ts"]
