+++ packages/tenor-api/package.json
@@ -1,0 +1,30 @@
+{
+  "name": "tenor-api",
+  "version": "1.0.0",
+  "type": "module",
+  "description": "Tenor API client for GIF search and trending",
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "scripts": {
+    "build": "tsc",
+    "test": "vitest run",
+    "test:run": "vitest run",
+    "lint": "eslint src --ext .ts"
+  },
+  "keywords": ["tenor", "api", "gif"],
+  "author": "",
+  "license": "MIT",
+  "dependencies": {},
+  "devDependencies": {
+    "typescript": "~5.8.3",
+    "vitest": "^2.0.0",
+    "eslint": "^9.25.0",
+    "eslint-config-expo": "~9.2.0",
+    "@eslint/eslintrc": "^3.0.0",
+    "@eslint/js": "^9.25.0"
+  },
+  "peerDependencies": {
+    "react": ">=18.0.0",
+    "react-native": ">=0.70.0"
+  }
+}
